+++ src/server/integration/sh/pre-process/routines.d/sld/correct-ogc-filter.sh
@@ -5,9 +5,11 @@ new_filepath=$filepath_noext-new
 if [ -n "$(cat $filepath | grep 'PropertyIsGreaterThan>')" ]; then
 	log "$filename"
 
-	cat $filepath | sed -e 's/PropertyIsGreaterThan>/PropertyIsGreaterThanOrEqualTo>/g' > $new_filepath.sld
+	if [[ $MODE = 'APPLY' ]]; then
+		cat $filepath | sed -e 's/PropertyIsGreaterThan>/PropertyIsGreaterThanOrEqualTo>/g' > $new_filepath.sld
 
-	rm $filepath
-	mv $new_filepath.sld $filepath
+		rm $filepath
+		mv $new_filepath.sld $filepath
+	fi
 
 fi
